+++ test/failurepolicy_test.go
@@ -2,6 +2,7 @@ package test
 
 import (
 	"errors"
+	"io"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -11,13 +12,25 @@ import (
 	"github.com/failsafe-go/failsafe-go/internal/testutil"
 )
 
-func TestHandleCustomErrorType(t *testing.T) {
+func TestHandleErrors(t *testing.T) {
 	fb := fallback.BuilderWithResult(true).
-		HandleErrors(&testutil.CompositeError{}).
+		HandleErrors(io.EOF).
 		Build()
 
 	result, err := failsafe.Get(func() (bool, error) {
-		return false, testutil.NewCompositeError(errors.New("test"))
+		return false, io.EOF
+	}, fb)
+	assert.True(t, result)
+	assert.Nil(t, err)
+}
+
+func TestHandleErrorsAs(t *testing.T) {
+	fb := fallback.BuilderWithResult(true).
+		HandleErrorTypes(testutil.CompositeError{}).
+		Build()
+
+	result, err := failsafe.Get(func() (bool, error) {
+		return false, testutil.CompositeError{Cause: errors.New("test")}
 	}, fb)
 	assert.True(t, result)
 	assert.Nil(t, err)